Renamed docs-config to swagger-config
